COMP: Clear warnings

See merge request iMSTK/iMSTK!420

+++ Source/Collision/CollisionDetection/imstkMeshToMeshBruteForceCD.cpp
@@ -42,6 +42,7 @@ MeshToMeshBruteForceCD::MeshToMeshBruteForceCD(std::shared_ptr<Geometry>      ob
 {
 }
 
+// The above error is caused by the 
 void
 MeshToMeshBruteForceCD::computeCollisionData()
 {
@@ -80,7 +81,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                                        m_proximityTolerance,
                                        m_proximityTolerance))
                 {
-                    m_colData->VTColData.safeAppend({ static_cast<unsigned int>(i), static_cast<unsigned int>(j), 0.0 });
+                    m_colData->VTColData.safeAppend({ static_cast<uint32_t>(i), static_cast<uint32_t>(j), 0.0 });
                 }
             }
         }
@@ -90,7 +91,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
         std::vector<std::vector<bool>> E2(numVertices, std::vector<bool>(numVertices, 1));
         const auto                     triangleVertices = mesh2->getTrianglesVertices();
 
-        for (int k = 0; k < numLines; ++k)
+        for (size_t k = 0; k < numLines; ++k)
         {
             const auto   nodes = mesh1->getLinesVertices()[k];
             const size_t i1    = nodes[0];
@@ -113,7 +114,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                         p0[0], p0[1], p0[2],
                         p1[0], p1[1], p1[2], m_proximityTolerance, m_proximityTolerance))
                     {
-                        m_colData->EEColData.safeAppend({ { i1, i2 }, { e[0], e[1] }, 0.0 });
+                        m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[0], (uint32_t)e[1] }, 0.0 });
                     }
                     E2[e[0]][e[1]] = 0;
                 }
@@ -125,7 +126,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                         p1[0], p1[1], p1[2],
                         p2[0], p2[1], p2[2], m_proximityTolerance, m_proximityTolerance))
                     {
-                        m_colData->EEColData.safeAppend({ { i1, i2 }, { e[1], e[2] }, 0.0 });
+                        m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[1], (uint32_t)e[2] }, 0.0 });
                     }
                     E2[e[1]][e[2]] = 0;
                 }
@@ -137,7 +138,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                         p2[0], p2[1], p2[2],
                         p0[0], p0[1], p0[2], m_proximityTolerance, m_proximityTolerance))
                     {
-                        m_colData->EEColData.safeAppend({ { i1, i2 }, { e[2], e[0] }, 0.0 });
+                        m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[2], (uint32_t)e[0] }, 0.0 });
                     }
                     E2[e[2]][e[0]] = 0;
                 }
@@ -167,7 +168,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                     p1[0], p1[1], p1[2],
                     p2[0], p2[1], p2[2], m_proximityTolerance, m_proximityTolerance))
                 {
-                    m_colData->VTColData.safeAppend({ static_cast<unsigned int>(i), static_cast<unsigned int>(j), 0.0 });
+                    m_colData->VTColData.safeAppend({ static_cast<uint32_t>(i), static_cast<uint32_t>(j), 0.0 });
                 }
             }
         }
@@ -195,7 +196,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                     p1[0], p1[1], p1[2],
                     p2[0], p2[1], p2[2], m_proximityTolerance, m_proximityTolerance))
                 {
-                    m_colData->VTColData.safeAppend({ static_cast<unsigned int>(i), static_cast<unsigned int>(j), 0.0 });
+                    m_colData->VTColData.safeAppend({ static_cast<uint32_t>(i), static_cast<uint32_t>(j), 0.0 });
                 }
             }
         }
@@ -212,8 +213,9 @@ MeshToMeshBruteForceCD::computeCollisionData()
         for (size_t k = 0; k < elements.size(); ++k)
         {
             const SurfaceMesh::TriangleArray& tri = elements[k];
-            size_t                            i1  = tri[0];
-            size_t                            i2  = tri[1];
+
+            size_t i1 = tri[0];
+            size_t i2 = tri[1];
 
             if (E[i1][i2] && E[i2][i1])
             {
@@ -234,7 +236,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p0[0], p0[1], p0[2],
                             p1[0], p1[1], p1[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[0], e[1] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[0], (uint32_t)e[1] }, 0.0 });
                             E2[e[0]][e[1]] = 0;
                         }
                     }
@@ -246,7 +248,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p1[0], p1[1], p1[2],
                             p2[0], p2[1], p2[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[1], e[2] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[1], (uint32_t)e[2] }, 0.0 });
                             E2[e[1]][e[2]] = 0;
                         }
                     }
@@ -258,7 +260,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p2[0], p2[1], p2[2],
                             p0[0], p0[1], p0[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[2], e[0] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[2], (uint32_t)e[0] }, 0.0 });
                             E2[e[2]][e[0]] = 0;
                         }
                     }
@@ -288,7 +290,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p0[0], p0[1], p0[2],
                             p1[0], p1[1], p1[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[0], e[1] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[0], (uint32_t)e[1] }, 0.0 });
                             E2[e[0]][e[1]] = 0;
                         }
                     }
@@ -300,7 +302,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p1[0], p1[1], p1[2],
                             p2[0], p2[1], p2[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[1], e[2] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[1], (uint32_t)e[2] }, 0.0 });
                             E2[e[1]][e[2]] = 0;
                         }
                     }
@@ -312,7 +314,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p2[0], p2[1], p2[2],
                             p0[0], p0[1], p0[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[2], e[0] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[2], (uint32_t)e[0] }, 0.0 });
                             E2[e[2]][e[0]] = 0;
                         }
                     }
@@ -341,7 +343,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p0[0], p0[1], p0[2],
                             p1[0], p1[1], p1[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[0], e[1] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[0], (uint32_t)e[1] }, 0.0 });
                             E2[e[0]][e[1]] = 0;
                         }
                     }
@@ -353,7 +355,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p1[0], p1[1], p1[2],
                             p2[0], p2[1], p2[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[1], e[2] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[1], (uint32_t)e[2] }, 0.0 });
                             E2[e[1]][e[2]] = 0;
                         }
                     }
@@ -365,7 +367,7 @@ MeshToMeshBruteForceCD::computeCollisionData()
                             p2[0], p2[1], p2[2],
                             p0[0], p0[1], p0[2], m_proximityTolerance, m_proximityTolerance))
                         {
-                            m_colData->EEColData.safeAppend({ { i1, i2 }, { e[2], e[0] }, 0.0 });
+                            m_colData->EEColData.safeAppend({ { (uint32_t)i1, (uint32_t)i2 }, { (uint32_t)e[2], (uint32_t)e[0] }, 0.0 });
                             E2[e[2]][e[0]] = 0;
                         }
                     }
